==> R/02.4-add_escolas_rais.R <==
source("R/fun/empregos/empregos.R")


# 4) Trazer informacoes de funcionarios de escolas publicas do censo escolar 
rais_bring_schools(2017)


# 4) Trazer informacoes de funcionarios de escolas publicas do censo escolar 
rais_bring_schools(2018)


# 4) Trazer informacoes de funcionarios de escolas publicas do censo escolar 
rais_bring_schools(2019)
